Add compiler warning level flag and resolve warnings
Most compile warnings were missing due to an automatically set compiler flag resulting in hidden warnings gathering over a long time and only now they've shown up and been fixed.

+++ pxlFramework/src/Core/Input.cpp
@@ -132,8 +132,8 @@ namespace pxl
         // TODO: check if x and y values are outside the bounds of the window
 
         glfwSetCursorPos(s_WindowHandle, x, y);
-        s_CursorPosition.x = x;
-        s_CursorPosition.y = y;
+        s_CursorPosition.x = static_cast<float>(x);
+        s_CursorPosition.y = static_cast<float>(y);
     }
 
     void Input::SetCursorMode(CursorMode cursorMode)
