+++ pxlFramework/src/Debug/ImGui/ImGuiVulkan.cpp
@@ -31,7 +31,7 @@ namespace pxl
 	    pool_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
 	    pool_info.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
 	    pool_info.maxSets = 1000;
-	    pool_info.poolSizeCount = std::size(pool_sizes);
+	    pool_info.poolSizeCount = static_cast<uint32_t>(std::size(pool_sizes));
 	    pool_info.pPoolSizes = pool_sizes;
 
 	    VK_CHECK(vkCreateDescriptorPool(m_Device->GetVkDevice(), &pool_info, nullptr, &m_ImGuiDescriptorPool));
